chore: rename `sips/dataclasses.py` to `sips/data.py` to avoid confusion with Python's `dataclasses`
